Shader class
Shader container class and shader manager singleton class.

Also attempted to make a templeted function that passes on a ellipsis parameter of pre-undefined types and lengths of parameters, but that is apparantly (not too surprisingly) not allowed. This of course makes me wonder how std::make_unique and std::make_shared manages to do this?