Added some kind of optimization technique
Used a std::set to make a sorted list for gameobjects which I then use to iterate through during rendering. This way the gameobjects are sorted by their shaders and I can limit the amount of shaderswitches and therefore also the amount of uniform data that needs to be sent to the shader, which is much more efficient. So the efficiency lost this way is because of the extra added sorting from std::set, but I won't need to sort more often then whenever I add or remove an element, which during runtime isn't that often.