Some letters on the screen
Proved that it works, but there's a better way of doing it described on https://learnopengl.com/In-Practice/Text-Rendering. If i make a quad that moves and scales when rendering each letter and also combine the individual textures for each letter into a texture atlas, I it can be both easier and faster to render text.